Added error_callback.h, Moved OpenGL code
