+++ Snake_game_cpp/error_callback.h
@@ -1,0 +1,2 @@
+// error_callback.h
+
